external functions try
